fix werror

+++ mlir/lib/IR/BuiltinTypes.cpp
@@ -566,8 +566,8 @@ unsigned MemRefType::getMemorySpaceAsInt() const {
   return detail::getMemorySpaceAsInt(getMemorySpace());
 }
 
-static int64_t getNumContiguousTrailingDimsImpl(ArrayRef<int64_t> shape,
-                                                ArrayRef<int64_t> strides) {
+static size_t getNumContiguousTrailingDimsImpl(ArrayRef<int64_t> shape,
+                                               ArrayRef<int64_t> strides) {
   const int64_t n = shape.size();
   // A memref with dimensions `d0, d1, ..., dn-1` and strides
   // `s0, s1, ..., sn-1` is contiguous up to dimension `k`
